Delegate Proposals Link
The delegate proposals on the 'forum' are usually dated. The most up to date proposals are on arkdelegates.io
The forum is also not attractive. In my opinion it looks unprofessional and unclean. It reminds me of a craigslist page. 
Also, saying the 'forum' is the 'official' delegates page implies centralization in a way. arkdelegates.io is maintained by an actual Ark Delegate.